+++ src/Screens/SampleScreen/__LC__/SampleWidgets.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useState } from 'react';
 import { useLocalSearchParams } from 'expo-router';
 
-import { GPS_DTO } from '@Types/ProjectTypes';
+import { GPS_DTO, WidgetData } from '@Types/ProjectTypes';
 import ProjectService from '@Services/ProjectService';
 import CacheService from '@Services/CacheService';
 import SyncService from '@Services/SyncService';
@@ -20,9 +20,11 @@ export function F_SampleWidgets(props: {
   const id_sample  = useLocalSearchParams().id_sample as string;
   const [_, refresh] = useState<boolean>(true);
 
-  const onDeleteWidget_Sample = useCallback(async (id_widget: string) => {
+  const onDeleteWidget_Sample = useCallback(async (widgetData: WidgetData) => {
+    const { id_widget } = widgetData;
     await ProjectService.deleteWidget_Sample(id_project, id_sample, id_widget,
       async () => {
+        ProjectService.deleteMedia_Widget(id_project, widgetData);
         CacheService.removeFromAllWidgets_Sample(id_widget);
         await SyncService.syncData_SampleWidgets(id_project, id_sample, id_widget, 'deletion');
         refresh(prev => !prev);
@@ -41,7 +43,7 @@ export function F_SampleWidgets(props: {
       }}
       widgetData={widgetData}
       referenceGPSData={props.referenceGPS}
-      onDeleteWidget={async () => await onDeleteWidget_Sample(widgetData.id_widget)}
+      onDeleteWidget={async () => await onDeleteWidget_Sample(widgetData)}
     />
   ));